Add variable_index.py
